backend/pyproject.toml
[project]
name = "stock-dashboard-backend"
version = "0.1.0"
description = "Stock Dashboard Backend — FastAPI + DuckDB"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "duckdb>=1.2.0",
    "yfinance>=0.2.40",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "apscheduler>=3.10",
    "anthropic>=0.86.0",
    "finnhub-python>=2.4",
    "httpx>=0.27",
    "pandas>=2.2",
    "numpy>=1.26",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["config", "db", "models", "services", "providers", "routers"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
